requirements.txt
fastapi
uvicorn
python-multipart
jinja2
python-docx
websockets
